+++ app/src/main/java/au/com/safetychampion/MainActivity.kt
@@ -13,14 +13,14 @@ import androidx.core.view.isVisible
 import androidx.lifecycle.lifecycleScope
 import au.com.safetychampion.data.domain.core.Result
 import au.com.safetychampion.databinding.ActivityMainBinding
+import au.com.safetychampion.util.asJson
 import au.com.safetychampion.utils.AssetsManager
-import au.com.safetychampion.utils.asJson
 import kotlinx.coroutines.flow.collectLatest
 import kotlinx.coroutines.launch
-import org.koin.android.ext.android.inject
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MainActivity : AppCompatActivity() {
-    private val viewModel by inject<MainViewModel>()
+    private val viewModel by viewModel<MainViewModel>()
     private val sampleData = AssetsManager(this)
     private val listUseCase = listOf(
         "Get Active Task (tasks/list/active)" to { viewModel.loadActiveTasks() },
