+++ app/src/main/java/au/com/safetychampion/di/AppModules.kt
@@ -1,9 +1,9 @@
 package au.com.safetychampion.di
 
 import au.com.safetychampion.MainViewModel
-import org.koin.core.module.dsl.singleOf
+import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
 
 val appModules = module {
-    singleOf(::MainViewModel)
+    viewModelOf(::MainViewModel)
 }
