+++ data/src/main/java/au/com/safetychampion/util/gsonadapter/BasePLTypeAdapter.kt
@@ -1,8 +1,8 @@
-package au.com.safetychampion.utils.gsonadapters
+package au.com.safetychampion.util.gsonadapter
 
-import au.com.safetychampion.data.di.retrofit.IGsonManager
 import au.com.safetychampion.data.domain.base.BasePL
-import au.com.safetychampion.utils.koinGet
+import au.com.safetychampion.util.IGsonManager
+import au.com.safetychampion.util.koinGet
 import com.google.gson.JsonElement
 import com.google.gson.JsonNull
 import com.google.gson.JsonSerializationContext
