+++ data/src/main/java/au/com/safetychampion/util/gsonadapter/TierTypeConverter.kt
@@ -1,4 +1,4 @@
-package au.com.safetychampion.utils.gsonadapters // ktlint-disable filename
+package au.com.safetychampion.util.gsonadapter // ktlint-disable filename
 
 import au.com.safetychampion.data.domain.models.TierType
 import com.google.gson.JsonDeserializationContext
